feat: разложил сервисы по пакетам и добавил авторизацию при запросе данных пользователя
