+++ src/main/kotlin/database/dao/AccountDAO.kt
@@ -2,7 +2,7 @@ package database.dao
 
 
 import database.tables.UsersTable
-import domain.Account
+import domain.user.Account
 import org.jetbrains.exposed.sql.*
 import org.jetbrains.exposed.sql.SqlExpressionBuilder.eq
 import org.jetbrains.exposed.sql.transactions.transaction
