+++ src/main/kotlin/domain/auth/AuthResult.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.auth
 
 /**
  * Результат аутентификации
