+++ src/test/kotlin/auth/AuthServiceTest.kt
@@ -1,8 +1,8 @@
 package auth
 
-import domain.Account
-import domain.AuthResult
-import domain.ResultCode
+import domain.user.Account
+import domain.auth.AuthResult
+import domain.auth.ResultCode
 import grpc.AuthProto.AuthResponse
 import grpc.AuthProto.LoginRequest
 import grpc.AuthProto.RegisterRequest
@@ -12,6 +12,8 @@ import io.mockk.verify
 import io.mockk.every
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import services.auth.AuthServiceImpl
+import services.auth.Authenticator
 
 class AuthServiceTest {
