+++ src/test/kotlin/auth/AuthenticatorTest.kt
@@ -2,13 +2,14 @@ package auth
 
 import database.dao.DAO.DatabaseException
 import database.manager.DatabaseManager
-import domain.Account
-import domain.ResultCode
+import domain.user.Account
+import domain.auth.ResultCode
 import io.mockk.every
 import io.mockk.mockk
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
+import services.auth.Authenticator
 import java.util.Optional
 
 class AuthenticatorTest {
