+++ src/test/kotlin/database/DbTableMockTest.kt
@@ -2,7 +2,7 @@ package database
 
 import database.dao.DAO
 import database.dao.DbTableMock
-import domain.Account
+import domain.user.Account
 import org.junit.jupiter.api.Assertions.*
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
